requirements.txt
django>=3
django-filter
djangorestframework
psycopg2-binary
pytelegrambotapi
